Fixed - issue in App.js and index.js
