+++ src/App.js
@@ -1,0 +1,11 @@
+import HomePage from './pages/HomePage/HomePage';
+
+function App() {
+  return (
+    <div className="App">
+      <HomePage />
+    </div>
+  );
+}
+
+export default App;
